File descriptors and permissions

+++ 0x15-file_io/3-cp.c
@@ -2,17 +2,18 @@
 
 /**
  * main - check the code
- *
+ * @ac: args num
+ * @av: args array
  * Return: Always 0.
  */
 int main(int ac, char **av)
 {
-	ssize_t source, destination, byteRead, byteWrite;
+	int source, destination, byteRead, byteWrite;
 	char buffer[1024];
-	
+
 	if (ac != 3)
 		exit_and_print("Usage: cp file_from file_to\n", av[0], 97);
-	
+
 	source = read_textfile(av[1], 114);
 	if (source == -1)
 		exit_and_print("Error: Can't read from file %s\n", av[1], 98);
@@ -32,10 +33,10 @@ int main(int ac, char **av)
 		exit_and_print("Error: Can't read from file %s\n", av[1], 98);
 
 	if (close(source) == -1)
-		exit_and_print("Error: Can't close fd %d\n", (int)source, 100);
+		exit_and_print("Error: Can't close fd %d\n", source, 100);
 
 	if (close(destination) == -1)
-		exit_and_print("Error: Can't close fd %d\n", (int)destination, 100);
+		exit_and_print("Error: Can't close fd %d\n", destination, 100);
 	return (0);
 
 }
